+++ lib/molinillo/delegates/resolution_state.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Molinillo
   # @!visibility private
   module Delegates
