+++ lib/molinillo/delegates/specification_provider.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Molinillo
   module Delegates
     # Delegates all {Molinillo::SpecificationProvider} methods to a
