+++ lib/molinillo/dependency_graph/action.rb
@@ -1,19 +1,35 @@
+# frozen_string_literal: true
 module Molinillo
   class DependencyGraph
+    # An action that modifies a {DependencyGraph} that is reversible.
+    # @abstract
     class Action
+      # rubocop:disable Lint/UnusedMethodArgument
+
+      # @return [Symbol] The name of the action.
       def self.name
         raise 'Abstract'
       end
 
-      def up(_graph)
+      # Performs the action on the given graph.
+      # @param  [DependencyGraph] graph the graph to perform the action on.
+      # @return [Void]
+      def up(graph)
         raise 'Abstract'
       end
 
-      def down(_graph)
+      # Reverses the action on the given graph.
+      # @param  [DependencyGraph] graph the graph to reverse the action on.
+      # @return [Void]
+      def down(graph)
         raise 'Abstract'
       end
 
-      attr_accessor :previous, :next
+      # @return [Action,Nil] The previous action
+      attr_accessor :previous
+
+      # @return [Action,Nil] The next action
+      attr_accessor :next
     end
   end
 end
