+++ lib/molinillo/dependency_graph/add_edge_no_circular.rb
@@ -1,14 +1,18 @@
+# frozen_string_literal: true
 require 'molinillo/dependency_graph/action'
 module Molinillo
   class DependencyGraph
     # @!visibility private
-    class AddEdgeNoCircular < Action # :nodoc:
+    # (see DependencyGraph#add_edge_no_circular)
+    class AddEdgeNoCircular < Action
       # @!group Action
 
+      # (see Action.name)
       def self.name
         :add_vertex
       end
 
+      # (see Action#up)
       def up(graph)
         edge = make_edge(graph)
         edge.origin.outgoing_edges << edge
@@ -16,6 +20,7 @@ module Molinillo
         edge
       end
 
+      # (see Action#down)
       def down(graph)
         edge = make_edge(graph)
         edge.origin.outgoing_edges.delete(edge)
@@ -24,14 +29,25 @@ module Molinillo
 
       # @!group AddEdgeNoCircular
 
+      # @return [String] the name of the origin of the edge
       attr_reader :origin
+
+      # @return [String] the name of the destination of the edge
       attr_reader :destination
+
+      # @return [Object] the requirement that the edge represents
       attr_reader :requirement
 
+      # @param  [DependencyGraph] graph the graph to find vertices from
+      # @return [Edge] The edge this action adds
       def make_edge(graph)
         Edge.new(graph.vertex_named(origin), graph.vertex_named(destination), requirement)
       end
 
+      # Initialize an action to add an edge to a dependency graph
+      # @param [String] origin the name of the origin of the edge
+      # @param [String] destination the name of the destination of the edge
+      # @param [Object] requirement the requirement that the edge represents
       def initialize(origin, destination, requirement)
         @origin = origin
         @destination = destination
