+++ lib/molinillo/dependency_graph/add_vertex.rb
@@ -1,14 +1,18 @@
+# frozen_string_literal: true
 require 'molinillo/dependency_graph/action'
 module Molinillo
   class DependencyGraph
     # @!visibility private
+    # (see DependencyGraph#add_vertex)
     class AddVertex < Action # :nodoc:
       # @!group Action
 
+      # (see Action.name)
       def self.name
         :add_vertex
       end
 
+      # (see Action#up)
       def up(graph)
         if existing = graph.vertices[name]
           @existing_payload = existing.payload
@@ -21,6 +25,7 @@ module Molinillo
         vertex
       end
 
+      # (see Action#down)
       def down(graph)
         if defined?(@existing_payload)
           vertex = graph.vertices[name]
@@ -33,10 +38,19 @@ module Molinillo
 
       # @!group AddVertex
 
+      # @return [String] the name of the vertex
       attr_reader :name
+
+      # @return [Object] the payload for the vertex
       attr_reader :payload
+
+      # @return [Boolean] whether the vertex is root or not
       attr_reader :root
 
+      # Initialize an action to add a vertex to a dependency graph
+      # @param [String] name the name of the vertex
+      # @param [Object] payload the payload for the vertex
+      # @param [Boolean] root whether the vertex is root or not
       def initialize(name, payload, root)
         @name = name
         @payload = payload
