+++ spec/dependency_graph/log_spec.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'spec_helper'
 
 describe Molinillo::DependencyGraph::Log do
