+++ spec/spec_helper.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'bundler/setup'
 
 # Set up coverage analysis
