+++ spec/spec_helper/index.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Molinillo
   class TestIndex
     attr_accessor :specs
