+++ spec/spec_helper/ui.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module Molinillo
   class TestUI
     include UI
