+++ spec/state_spec.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require File.expand_path('../spec_helper', __FILE__)
 
 module Molinillo
